+++ fsleyes/displaycontext/canvasopts.py
@@ -171,17 +171,42 @@ class LightBoxCanvasOpts(SliceCanvasOpts):
     The Z axis of the display coordinate system (controlled by the
     :attr:`zax` property) is divided up into slices, which are defined
     in a "slice coordinate system" having range 0 to 1, according to the
-    :attr:`zrange` and :attr:`sliceSpacing` properties::
+    :attr:`zrange` and :attr:`sliceSpacing` properties.
 
-                     zlo                                     zhi
+    Two sampling regimes are supported, controlled by the :attr:`sampleSlices`
+    property.  In the first regime (``sampleSlices = 'centre'``), the entire
+    0-1 range is divided into N slices, and each slice is sampled at its
+    centre. The slices within which the low and high ``zrange`` bound are
+    located determine the first and last slices to be displayed::
+
+                           zlo                     zhi
+                            |                       |
                       0                                       1
         slice centre  |   *       *       *       *       *   |
         slice index   |   0   |   1   |   2   |   3   |   4   |
                       |-------|
                        spacing
 
-    where ``*`` denotes the slice centres - these correspond to the
-    Z positions returned by the :meth:`slices` property.
+    where ``*`` denotes the slice sampling points - these correspond to the Z
+    positions returned by the :meth:`slices` property.
+
+    In the second regime (``sampleSlices = 'start'``), the 0-1 range is
+    divided into N slices such that the low ``zrange`` bound falls at the
+    beginning of a slice. Each slice is sampled at its beginning::
+
+                          zlo                        zhi
+                           |                          |
+                      0                                       1
+        slice start   |    |*      |*      |*      |*      |* |
+        slice index   |  0 |   1   |   2   |   3   |   4   |  |
+                           |-------|
+                            spacing
+
+    The first ``'centre'`` regime is the default, and is useful when viewing
+    overlays with different dimensions, where the specific slices being
+    displayed are not important. The second ``'start'`` regime is useful when a
+    specific starting slice is desired. See the :class:`.LightBoxSamplePanel`
+    for an example using the :attr:`sampleSlices` property.
     """
 
 
@@ -299,6 +324,14 @@ class LightBoxCanvasOpts(SliceCanvasOpts):
     """
 
 
+    sampleSlices = props.Choice(('centre', 'start'))
+    """Controls whether the slice positions returned by :meth:`slices`
+    correspond to slice centres or to slice starts. This is used by the
+    :class:`.LightBoxSampleDialog`, which allows the user to select slices
+    in terms of voxel coordinates.
+    """
+
+
     def __init__(self):
         super().__init__()
         name = self.name
@@ -375,10 +408,10 @@ class LightBoxCanvasOpts(SliceCanvasOpts):
 
     @property
     def nslices(self):
-        """Returns the total number of slices that should currently be displayed,
-        i.e. that fit within the current :attr:`zrange`. This is automatically
-        calculated from the current :attr:`zrange` and :attr:`sliceSpacing`
-        settings.
+        """Returns the total number of slices that should currently be
+        displayed, i.e. that fit within the current :attr:`zrange`. This is
+        automatically calculated from the current :attr:`zrange` and
+        :attr:`sliceSpacing` settings.
 
         Note that this may be different to the :meth:`.LightBoxCanvas.nslices`,
         which returns the total number of slices that can be displayed on the
@@ -405,8 +438,27 @@ class LightBoxCanvasOpts(SliceCanvasOpts):
         the :meth:`startslice` to identify the index of the starting slice to
         be displayed. The locations correspond to slice centres.
         """
-        sliceSpacing = self.sliceSpacing
-        return np.arange(sliceSpacing / 2, 1, sliceSpacing)
+        ssp = self.sliceSpacing
+
+        # ssp / 2 -> sampling from slice centres
+        if self.sampleSlices == 'centre':
+            start = ssp / 2
+
+        # Sample from slice beginning. Start slices
+        # at a location such that zlo is located
+        # at the beginning of a slice.
+
+        # Add a small fixed offset to ensure that
+        # we're within the region corresponding to
+        # each slice, as otherwise e.g. the first
+        # two slices [0.0, <sliceSpacing>] may
+        # resolve to the same slice.
+        else:
+            zlo    = self.zrange.xlo
+            start  = zlo - ssp * np.floor(zlo / ssp)
+            start += 0.0001
+
+        return np.arange(start, 1, ssp)
 
 
     @property